Refactor activity related code from robot.* to activities.*
